sysdeps: look in C{,PLUS}_INCLUDE_PATH for headers
jhbuild itself sets this variable, but it doesn't seem to consult it
when looking for sysdeps.

For additional system library paths, it turns out that setting this
variable is a better idea than using -I or -isystem in CFLAGS because
those options are "more powerful" than C_INCLUDE_PATH which is what
jhbuild uses for pointing to the include directory in its prefix (which
should take precedence).

https://bugzilla.gnome.org/show_bug.cgi?id=725986