config: Add an 'extra_prefixes' variable
Add an extra_prefixes variable that will instruct jhbuild to act as it
has been run inside of itself with two separate prefixes.  This allows
setting up a 'cascading' environment where you could have multiple
layers of jhbuild on top of each other (for different sets of work).

It also allows for proper handling of /usr/local on systems that install
software there (and we set 'extra_prefixes' to contain '/usr/local' by
default on those systems).

The extra_prefixes list is traversed so that the first entry has highest
precedence.