Revamp PYTHONPATH setup
jhbuild currently tries, at runtime, to detect the version of python in
use and guess an appropriate setting for the PYTHONPATH.  This variable
ends up containing a string like '.../lib/python2.7/site-packages' which
is a problem, since it is shared between python2 and python3.

We can work around this by setting PYTHONPATH to a directory shipped
with jhbuild that contains a sitecustomize hook that will dynamically
configure the correct path at runtime, depending on the newly-added
JHBUILD_PREFIXES environment variable.